+++ CHANGELOG.md
@@ -10,6 +10,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 ### **Added**
+
+- added `ray-on-eks`, and `manifests/ray-on-eks` manifests
+- Added a `sagemaker-model-monitoring-module` module with an example of data quality, model quality, model bias, and model explainability monitoring of a SageMaker Endpoint.
+- Added an option to enable data capture in the `sagemaker-endpoint-module`.
+
+### **Changed**
+- remove explicit module manifest account/region mappings from `fmops-qna-rag`
+- changed ECR encryption to KMS_MANAGED
+- changed encryption for each bucket to KMS_MANAGED
+- refactor `airflow-dags` module to use Pydantic
+- fix inputs for `bedrock-finetuning` module not working
+- add `retention-type` argument for the bucket in the `bedrock-finetuning` module
+- fix broken dependencies for `examples/airflow-dags`
+
+## v1.2.0
+
+### **Added**
 - added multi-acc sagemaker-mlops manifest example
 
 ### **Changed**
@@ -21,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `sagemaker-studio`
   - `sagemaker-endpoint`
   - `sagemaker-templates-service-catalog`
+  - `sagemaker-custom-kernel`
   - `qna-rag`
 - add CDK nag to `qna-rag` module
 - rename seedfarmer project name to `aiops`
@@ -46,6 +64,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added `sagemaker-hugging-face-endpoint` module
 - added `hf_import_models` template to import hugging face models
 - added `qna-rag` module
+- added `bedrock-finetuning` module
 
 ### **Changed**